ref(github): Move GITHUB_API_ACCEPT_HEADER to constants module
Resolve circular imports between integration.py and client.py /
provider.py by moving the shared constant to the lightweight
constants.py module.
