+++ src/sentry/identity/github/provider.py
@@ -2,11 +2,12 @@ from django.core.exceptions import PermissionDenied
 
 from sentry import http, options
 from sentry.identity.oauth2 import OAuth2Provider
-from sentry.integrations.github.integration import GITHUB_API_ACCEPT_HEADER
 from sentry.integrations.types import IntegrationProviderSlug
 
 
 def get_user_info(access_token):
+    from sentry.integrations.github.constants import GITHUB_API_ACCEPT_HEADER
+
     with http.build_session() as session:
         resp = session.get(
             "https://api.github.com/user",
