+++ src/sentry/integrations/github/constants.py
@@ -1,3 +1,5 @@
+GITHUB_API_ACCEPT_HEADER = "application/vnd.github+json"
+
 ISSUE_LOCKED_ERROR_MESSAGE = "Unable to create comment because issue is locked."
 
 RATE_LIMITED_MESSAGE = "API rate limit exceeded"
