+++ tests/sentry/integrations/github/test_client.py
@@ -14,7 +14,8 @@ from responses import matchers
 from sentry.constants import ObjectStatus
 from sentry.integrations.github.blame import create_blame_query, generate_file_path_mapping
 from sentry.integrations.github.client import GitHubApiClient, GitHubReaction
-from sentry.integrations.github.integration import GITHUB_API_ACCEPT_HEADER, GitHubIntegration
+from sentry.integrations.github.constants import GITHUB_API_ACCEPT_HEADER
+from sentry.integrations.github.integration import GitHubIntegration
 from sentry.integrations.source_code_management.commit_context import (
     CommitInfo,
     FileBlameInfo,
